+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased][unreleased]
+### Bugfix
+* Fixed bugs with parallel mode and performing param injection
 
 ## [3.0.0-rc.11][2016-08-30]
 ### Bugfix