+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased][unreleased]
+
+## [3.0.0-rc.13][2016-09-01]
 ### Bugfix
 * Fixed some more bugs with parallel mode